feat: accounting route
